"Usher forth 'server-info' to replace 'environment'—a new harbinger of deploy wisdom, streamlined and severed from its former abode."
