+++ api/tests/mocha/services/deploy-info.spec.js
@@ -4,7 +4,7 @@ const rewire = require('rewire');
 
 const fs = require('fs');
 
-const environment = require('@medic/environment');
+const serverInfo = require('@medic/server-info');
 const resources = require('../../../src/resources');
 
 let service;
@@ -29,25 +29,25 @@ describe('deploy info', () => {
         time: 'human readable',
       };
 
-      sinon.stub(environment, 'getDeployInfo').resolves(deployInfo);
+      sinon.stub(serverInfo, 'getDeployInfo').resolves(deployInfo);
 
       const result = await service.get();
 
       expect(result).to.deep.equal(deployInfo);
-      expect(environment.getDeployInfo.callCount).to.equal(1);
+      expect(serverInfo.getDeployInfo.callCount).to.equal(1);
     });
 
     it('should work with undefined deploy info and build info', async () => {
       const deployInfo = { version: '20000' };
 
-      sinon.stub(environment, 'getDeployInfo').resolves(deployInfo);
+      sinon.stub(serverInfo, 'getDeployInfo').resolves(deployInfo);
 
       expect(await service.get()).to.deep.equal(deployInfo);
     });
 
     it('should throw error when getDeployInfo fails', async () => {
       const error = { error: 'whatever' };
-      sinon.stub(environment, 'getDeployInfo').rejects(error);
+      sinon.stub(serverInfo, 'getDeployInfo').rejects(error);
 
       try {
         await service.get();
@@ -71,7 +71,7 @@ describe('deploy info', () => {
         version: '4.10.0-beta.1',
       };
 
-      sinon.stub(environment, 'getDeployInfo').resolves(deployInfo);
+      sinon.stub(serverInfo, 'getDeployInfo').resolves(deployInfo);
 
       await service.store();
