+++ shared-libs/server-info/package.json
@@ -1,0 +1,15 @@
+{
+  "name": "@medic/server-info",
+  "version": "1.0.0",
+  "description": "Server info utility functions",
+  "main": "src/index.js",
+  "scripts": {
+    "test": "nyc --nycrcPath='../nyc.config.js' mocha ./test"
+  },
+  "author": "",
+  "license": "Apache-2.0",
+  "dependencies": {
+    "@medic/environment": "file:../environment",
+    "@medic/couch-request": "file:../couch-request"
+  }
+}
